FEAT: Implementa a coleta de estatísticas de commits para GitHub e GitLab Signed-off-by: jadson santos <jadsonjs@gmail.com>

+++ src/main/java/br/com/jadson/snooper/github/data/issue/GitHubIssueInfo.java
@@ -52,9 +52,9 @@ public class GitHubIssueInfo {
     public String comments_url;
     public String events_url;
     public String html_url;
-    public int id;
+    public long id;
     public String node_id;
-    public int number;
+    public long number;
     public String title;
     public String body;
     public GitHubIssueUserInfo user;
